% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{add_logo}
\usage{
add_logo(plot_path, logo_position, logo_scale = 10)
}
\arguments{
\item{plot_path}{PARAM_DESCRIPTION}

\item{logo_position}{PARAM_DESCRIPTION}

\item{logo_scale}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
add ACDI/VOCA logo
}
