% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_AV.R
\name{theme_AVgray}
\alias{theme_AVgray}
\title{FUNCTION_TITLE}
\usage{
theme_AVgray(text.font = "Calibri", title.font = "Calibri",
  legend.font = "Calibri", title.size = 18, text.size = 14,
  subtitle.size = 12, axis.title.size = 14, axis.text.size = 12,
  legend.title.size = 10, legend.text.size = 9,
  title.color = "#F9FEFF", subtitle.color = "#F9FEFF",
  text.color = "#F9FEFF", axis.title.color = "#F9FEFF",
  axis.text.color = "#F9FEFF", legend.title.color = "#F9FEFF",
  legend.text.color = "#F9FEFF", legend.position = "bottom",
  ticks = FALSE)
}
\arguments{
\item{text.font}{PARAM_DESCRIPTION, Default: "Calibri"}

\item{title.font}{PARAM_DESCRIPTION, Default: "Calibri"}

\item{legend.font}{PARAM_DESCRIPTION, Default: "Calibri"}

\item{title.size}{PARAM_DESCRIPTION, Default: 18}

\item{text.size}{PARAM_DESCRIPTION, Default: 14}

\item{subtitle.size}{PARAM_DESCRIPTION, Default: 12}

\item{axis.title.size}{PARAM_DESCRIPTION, Default: 14}

\item{axis.text.size}{PARAM_DESCRIPTION, Default: 12}

\item{legend.title.size}{PARAM_DESCRIPTION, Default: 10}

\item{legend.text.size}{PARAM_DESCRIPTION, Default: 9}

\item{title.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{subtitle.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{text.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{axis.title.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{axis.text.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{legend.title.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{legend.text.color}{PARAM_DESCRIPTION, Default: '#F9FEFF'}

\item{legend.position}{PARAM_DESCRIPTION, Default: 'bottom'}

\item{ticks}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{margin}}
 \code{\link[grid]{unit}}
}
